third_party/blink/web_tests/platform/mac-mac11/virtual/disable-intersection-optimization/external/wpt/html/semantics/embedded-content/the-img-element/srcset/avoid-reload-on-resize-expected.txt

This is a testharness.js-based test.
[FAIL] Avoid srcset image reloads when viewport resizes
  Uncaught Error: assert_equals: expected 1 but got 2
Harness: the test ran to completion.
